Add default config options
Parsing config options is specific to pacman so im not realy sure if
this is should really be in the go-alpm code. Seems as its here already
I have extended it to set the same defaults as pacman.

Even if we do want it here or moved elsewhere the parsing code is kinda
a mess, I myself would prefer a more explicit method over reflection and
you can quite easily trigger segfaults by butting valid [options]
settings into a [repo] setting. This should probably get a rewrite at
some point.