Add support for question callback
Not all callback types are implemented as that would require defining
a lot of different types. The main purpose of this is to use the
ignorepkg question. If others are needed they can be added later.

The existing Question enmum and its values have all been named to
QuestionType, anything using these values will need to change them to
the name Name.