Fix searching for packages and groups
Earlier commits broke the error checking for these functions. There is
actually no need for error checking on these functions as pacman nevers
sets alpm_errno if a package is not found.

Also add a len and empty function to to clients can check if a search
has any results.